users/models.py
from django.contrib.auth.models import AbstractUser, UserManager
from phonenumber_field.modelfields import PhoneNumberField


class User(AbstractUser):
    phone_number = PhoneNumberField(verbose_name='Телефон', region='RU')

    objects = UserManager()
