requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-phonenumber-field==5.0.0
idna==2.10
phonenumberslite==8.12.20
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
